ADD SIDEBAR SECTION TO LAYOUT
While creating the side bar layout, I learned that float will not
work because when page is resized, floating elements would
automatically be repositioned to a newline if remaining width
is less than the floated element. I realized also that google's
search results is not responsive. So I decided to use absolute
positioning as the method of layout.