FINISH STYLING RELATED INFO SIDE BAR SECTION
One of the difficulties I encountered was with lining up
multiple image links next to each other. The solution I ended up
using was with inline-blocks. With a few hacks, such as setting
font-size to 0, setting the text-align, and vertical align
property,I was able to get the inline image links to behave the
way I wanted to.